requirements.txt
streamlit>=1.31.0
langchain>=0.1.0
langchain-community>=0.0.20
chromadb>=0.4.22
sentence-transformers>=2.2.2
ollama>=0.1.7
PyMuPDF>=1.23.0
python-docx>=1.1.0
tiktoken>=0.5.2
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.66.0
